2025-2029/Adithyanandan/CivicSense-AI/requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
autocommand==2.2.2
banks==2.4.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.4.0
distro==1.9.0
editables==0.5
filelock==3.24.0
filetype==1.2.0
Flask==3.1.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gmpy2==2.3.0
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
groq==1.0.0
h11==0.16.0
hatch==1.16.3
hatch-vcs==0.5.0
hatchling==1.28.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyperlink==21.0.0
idna==3.11
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
keyring==25.7.0
kiwisolver==1.4.7
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.15
llama-index-cli==0.5.4
llama-index-core==0.14.15
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.2
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-groq==0.4.1
llama-index-llms-openai==0.6.25
llama-index-llms-openai-like==0.5.3
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-postgres==0.7.3
llama-index-workflows==2.15.0
llama-parse==0.6.54
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
materialyoucolor==3.0.0
matplotlib==3.10.8
mdurl==0.1.2
meson==1.10.1
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.2
openai==2.24.0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pexpect==4.9.0
pgvector==0.4.2
pillow==12.1.1
pkg_resources==81.0.0
platformdirs==4.9.3
pluggy==1.6.0
pooch==1.9.0
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.54.5
pyparsing==3.3.2
pypdf==6.7.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
SecretStorage==3.3.3
sentence-transformers==5.2.3
sentencepiece==0.2.1
setuptools==82.0.0
setuptools-scm==9.2.2
shellingham==1.5.4
six==1.17.0
smbus==1.1
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
striprtf==0.0.26
sympy==1.13.3
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.0
tokenizers==0.22.2
tomli_w==1.2.0
tomlkit==0.14.0
torch==2.10.0+cpu
tqdm==4.67.3
transformers==4.57.6
trash-cli==0.24.5.26
trove-classifiers==2026.1.14.14
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
userpath==1.9.2
uv==0.10.6
virtualenv==20.39.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==1.17.3
yarl==1.23.0
